moving first things to the status bar
